refactor: just use "V2" standalone files
RHCLOUD-36297
