--- backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/drawer/DrawerResourceV1.java
@@ -1,9 +1,0 @@
-package com.redhat.cloud.notifications.routers.handlers.drawer;
-
-import jakarta.ws.rs.Path;
-
-import static com.redhat.cloud.notifications.Constants.API_NOTIFICATIONS_V_1_0;
-
-@Path(API_NOTIFICATIONS_V_1_0 + "/notifications/drawer")
-public class DrawerResourceV1 extends DrawerResource {
-}
