+++ backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/event/EventResource.java
@@ -23,6 +23,7 @@ import jakarta.validation.Valid;
 import jakarta.ws.rs.BadRequestException;
 import jakarta.ws.rs.BeanParam;
 import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.SecurityContext;
@@ -41,6 +42,7 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import static com.redhat.cloud.notifications.Constants.API_NOTIFICATIONS_V_1_0;
 import static com.redhat.cloud.notifications.auth.ConsoleIdentityProvider.RBAC_READ_NOTIFICATIONS_EVENTS;
 import static com.redhat.cloud.notifications.routers.SecurityContextUtil.getOrgId;
 import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
@@ -57,6 +59,11 @@ public class EventResource {
 
     @Inject
     WorkspaceUtils workspaceUtils;
+
+    @Path(API_NOTIFICATIONS_V_1_0 + "/notifications/events")
+    public static class V1 extends EventResource {
+
+    }
 
     @GET
     @Produces(APPLICATION_JSON)
