+++ backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/notification/NotificationResource.java
@@ -1,5 +1,6 @@
 package com.redhat.cloud.notifications.routers.handlers.notification;
 
+import com.redhat.cloud.notifications.Constants;
 import com.redhat.cloud.notifications.auth.ConsoleIdentityProvider;
 import com.redhat.cloud.notifications.auth.kessel.KesselAuthorization;
 import com.redhat.cloud.notifications.auth.kessel.permission.IntegrationPermission;
@@ -109,6 +110,40 @@ public class NotificationResource {
 
     @Inject
     EventTypeRepository eventTypeRepository;
+
+    @Path(Constants.API_NOTIFICATIONS_V_1_0 + "/notifications")
+    public static class V1 extends NotificationResource {
+        @GET
+        @Path("/eventTypes/{eventTypeId}/behaviorGroups")
+        @Produces(APPLICATION_JSON)
+        @Operation(summary = "List the behavior groups linked to an event type", description = "Lists the behavior groups that are linked to an event type. Use this endpoint to see which behavior groups will be affected if you delete an event type.")
+        public List<BehaviorGroup> getLinkedBehaviorGroups(
+            @Context SecurityContext sec,
+            @PathParam("eventTypeId") UUID eventTypeId,
+            @BeanParam @Valid Query query
+        ) {
+            if (this.backendConfig.isKesselRelationsEnabled(getOrgId(sec))) {
+                final UUID workspaceId = this.workspaceUtils.getDefaultWorkspaceId(getOrgId(sec));
+                this.kesselAuthorization.hasPermissionOnWorkspace(sec, WorkspacePermission.EVENT_TYPES_VIEW, workspaceId);
+                this.kesselAuthorization.hasPermissionOnWorkspace(sec, WorkspacePermission.BEHAVIOR_GROUPS_VIEW, workspaceId);
+
+                return this.internalGetLinkedBehaviorGroups(sec, eventTypeId, query);
+            } else {
+                return this.legacyRBACGetLinkedBehaviorGroups(sec, eventTypeId, query);
+            }
+        }
+
+        @RolesAllowed(ConsoleIdentityProvider.RBAC_READ_NOTIFICATIONS)
+        public List<BehaviorGroup> legacyRBACGetLinkedBehaviorGroups(final SecurityContext securityContext, final UUID eventTypeId, @Valid final Query query) {
+            return this.internalGetLinkedBehaviorGroups(securityContext, eventTypeId, query);
+        }
+
+        public List<BehaviorGroup> internalGetLinkedBehaviorGroups(final SecurityContext securityContext, final UUID eventTypeId, @Valid final Query query) {
+            String orgId = getOrgId(securityContext);
+
+            return behaviorGroupRepository.findBehaviorGroupsByEventTypeId(orgId, eventTypeId, query);
+        }
+    }
 
     @GET
     @Path("/eventTypes")
