--- backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/orgconfig/OrgConfigResourceV1.java
@@ -1,8 +1,0 @@
-package com.redhat.cloud.notifications.routers.handlers.orgconfig;
-
-import com.redhat.cloud.notifications.Constants;
-import jakarta.ws.rs.Path;
-
-@Path(Constants.API_NOTIFICATIONS_V_1_0 + "/org-config")
-public class OrgConfigResourceV1 extends OrgConfigResource {
-}
