--- backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/status/StatusResourceV1.java
@@ -1,9 +1,0 @@
-package com.redhat.cloud.notifications.routers.handlers.status;
-
-import jakarta.ws.rs.Path;
-
-import static com.redhat.cloud.notifications.Constants.API_NOTIFICATIONS_V_1_0;
-
-@Path(API_NOTIFICATIONS_V_1_0 + "/status")
-public class StatusResourceV1 extends StatusResource {
-}
