--- backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/userconfig/UserConfigResourceV1.java
@@ -1,8 +1,0 @@
-package com.redhat.cloud.notifications.routers.handlers.userconfig;
-
-import com.redhat.cloud.notifications.Constants;
-import jakarta.ws.rs.Path;
-
-@Path(Constants.API_NOTIFICATIONS_V_1_0 + "/user-config")
-public class UserConfigResourceV1 extends UserConfigResource {
-}
