--- backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/userpreferencespolicy/UserPreferencesForPolicyV1.java
@@ -1,8 +1,0 @@
-package com.redhat.cloud.notifications.routers.handlers.userpreferencespolicy;
-
-import com.redhat.cloud.notifications.Constants;
-import jakarta.ws.rs.Path;
-
-@Path(Constants.API_NOTIFICATIONS_V_1_0 + "/user-config")
-public class UserPreferencesForPolicyV1 extends UserPreferencesForPolicy {
-}
